pyproject.toml
[project]
name = "sample-backend-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask>=3.0.0",
    "openai>=1.53.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.1",
    "pypdf2>=3.0.1",
]